+++ lib/zaikio/mission_control.rb
@@ -84,6 +84,8 @@ require "zaikio/mission_control/desired_substrate"
 require "zaikio/mission_control/finishing_application"
 require "zaikio/mission_control/finishing"
 require "zaikio/mission_control/file_reference"
+require "zaikio/mission_control/machine"
+require "zaikio/mission_control/slot"
 
 module Zaikio
   module MissionControl