+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.5]
+
 - Add `Machine` and `Slot` resources, basic api retrieval for `Machine` production_schedule
 - Remove availabilities for machines and specialists (will be moved to the hub)
 
@@ -36,7 +38,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add finishings specification
 - Add jobs and parts specification
 
-[Unreleased]: https://github.com/zaikio/zaikio-mission_control-ruby/compare/v0.2.4..HEAD
+[Unreleased]: https://github.com/zaikio/zaikio-mission_control-ruby/compare/v0.2.5..HEAD
+[0.2.4]: https://github.com/zaikio/zaikio-mission_control-ruby/compare/v0.2.4..v0.2.5
 [0.2.4]: https://github.com/zaikio/zaikio-mission_control-ruby/compare/v0.2.3..v0.2.4
 [0.2.3]: https://github.com/zaikio/zaikio-mission_control-ruby/compare/v0.2.2..v0.2.3
 [0.2.2]: https://github.com/zaikio/zaikio-mission_control-ruby/compare/v0.2.1..v0.2.2